+++ testes/testeUtils.h
@@ -3,4 +3,4 @@
 #include <sys/time.h>
 #include <assert.h>
 
-int runTimedSort(void (*sorting_algorithm)(int *vet, int tam), int *vet, int tam);
+double runTimedSort(void (*sorting_algorithm)(int *vet, int tam), int *vet, int tam);
